discovery/kubernetes: extract pod discovery
This change extracts pod discovery into its own type.
